Extract "Document Preparer" from Elasticsearch::Index
DocumentPreparer _prepares_ the document before it can be sent to ES.
It is not in the Elasticsearch namespace because it has nothing to do
with Elasticsearch.

It causes some minor duplication which will be solved by improving
BestBetsChecker.
